# backend/app/api/v1/feed.py
"""``POST /api/v1/feed/recreate`` -- rotate the feed + episode GUIDs.

API-only (deliberately not surfaced in the UI). Requires an explicit
``confirm=true`` query parameter, mirroring ``/purge``: rotating the channel
``podcast:guid`` signals aggregators this is a new feed, and salting every
episode ``<guid>`` makes podcast apps re-download all episodes. Intended for
troubleshooting clients stuck on a stale, long-cached feed; disruptive to
existing subscribers, hence the confirm gate.
"""

from __future__ import annotations

import sqlite3
from typing import Annotated

from fastapi import APIRouter, Depends, HTTPException, Query
from pydantic import BaseModel, ConfigDict, Field

from app.api.deps import get_conn
from app.config import Settings, get_settings
from app.services import feed_revision, settings_store

router = APIRouter(tags=["maintenance"])


class FeedRecreateResponse(BaseModel):
    model_config = ConfigDict(extra="forbid")

    podcast_guid: str = Field(description="The new channel-level podcast:guid")
    guid_epoch: int = Field(description="Counter now appended to every episode <guid>")


@router.post(
    "/feed/recreate",
    response_model=FeedRecreateResponse,
)
async def post_feed_recreate(
    conn: Annotated[sqlite3.Connection, Depends(get_conn)],
    settings: Annotated[Settings, Depends(get_settings)],
    confirm: Annotated[
        bool,
        Query(description="Must be true to acknowledge the disruptive action."),
    ] = False,
) -> FeedRecreateResponse:
    if not confirm:
        raise HTTPException(
            status_code=400,
            detail="confirm=true is required to acknowledge the disruptive action",
        )
    conn.execute("BEGIN IMMEDIATE")
    try:
        guid, epoch = settings_store.rotate_feed_guids(conn, settings.BASE_URL, commit=False)
        feed_revision.bump(conn)
        conn.commit()
    except Exception:
        conn.rollback()
        raise
    return FeedRecreateResponse(podcast_guid=guid, guid_epoch=epoch)