# openapi.yaml
openapi: 3.1.0
info:
  title: Audicle
  version: 0.57.0
paths:
...
  /api/v1/feed/recreate:
    post:
      tags:
      - maintenance
      summary: Post Feed Recreate
      operationId: post_feed_recreate_api_v1_feed_recreate_post
      parameters:
      - name: confirm
        in: query
        required: false
        schema:
          type: boolean
          description: Must be true to acknowledge the disruptive action.
          default: false
          title: Confirm
        description: Must be true to acknowledge the disruptive action.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedRecreateResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
  /api/v1/feed-auth:
    get:
      tags:
      - feed-auth
      summary: Get Feed Auth
      operationId: get_feed_auth_api_v1_feed_auth_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedAuthStatus'
      security:
      - {}
      - AdminSession: []
    post:
      tags:
      - feed-auth
      summary: Set Feed Auth Enabled
      description: 'Enable or disable authenticated feeds.


        Enabling with no stored key mints one (retained across a later disable, so

        re-enabling reuses the same key rather than breaking every already-updated

        app again). ``set_if_absent`` is atomic, so two concurrent enables agree on

        one key instead of racing to different values -- the status then returns the

        key the operator can actually subscribe with. ``body.enabled`` is a real bool

        via Pydantic, so a stringly-typed ``"false"`` can''t slip through as True.'
      operationId: set_feed_auth_enabled_api_v1_feed_auth_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedAuthToggle'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedAuthStatus'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
  /api/v1/feed-auth/regenerate:
    post:
      tags:
      - feed-auth
      summary: Regenerate Feed Auth Key
      description: 'Rotate the key. 409 while disabled -- rotating a dormant key is
        invisible

        and surprising on re-enable. The old key stops working the moment the new
        one

        is stored, so every subscribed app must re-subscribe with the new URL.'
      operationId: regenerate_feed_auth_key_api_v1_feed_auth_regenerate_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedAuthStatus'
      security:
      - {}
      - AdminSession: []
        CsrfToken: []
...
components:
  schemas:
...
    FeedAuthStatus:
      properties:
        enabled:
          type: boolean
          title: Enabled
        key:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
        keyed_feed_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyed Feed Url
      additionalProperties: false
      type: object
      required:
      - enabled
      - key
      - keyed_feed_url
      title: FeedAuthStatus
    FeedAuthToggle:
      properties:
        enabled:
          type: boolean
          title: Enabled
      additionalProperties: false
      type: object
      required:
      - enabled
      title: FeedAuthToggle
    FeedRecreateResponse:
      properties:
        podcast_guid:
          type: string
          title: Podcast Guid
          description: The new channel-level podcast:guid
        guid_epoch:
          type: integer
          title: Guid Epoch
          description: Counter now appended to every episode <guid>
      additionalProperties: false
      type: object
      required:
      - podcast_guid
      - guid_epoch
      title: FeedRecreateResponse
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - status
      properties:
        error:
          type: string
        status:
          type: integer
        details:
          type: object
          additionalProperties: true
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: audicle_session
      description: Required for the admin API when a password is configured.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the admin session for unsafe methods.
    FeedKey:
      type: apiKey
      in: query
      name: key
      description: Required for protected feeds and non-artwork media.